Working on the vivado hls console and its quirks Also working towards factorizing the console structure to unify vivado, vivado and modelsim on a base class See #69
